Imported Upstream version 2.5.7

+++ timsieved/actions.c
@@ -39,45 +39,36 @@
  * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN
  * AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING
  * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
- *
- * $Id: actions.c,v 1.48 2010/01/06 17:02:01 murch Exp $
  */
 
 #ifdef HAVE_CONFIG_H
 #include <config.h>
 #endif
 
+#include <dirent.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <limits.h>
 #include <stdlib.h>
 #include <stdio.h>
-#include <sys/stat.h>
-#include <errno.h>
-#include <stdlib.h>
-#include <limits.h>
-#include <sys/param.h>
-#include <sys/types.h>
-#include <fcntl.h>
-#include <syslog.h>
-#include <dirent.h>
-#include <ctype.h>
-#include <unistd.h>
 #include <string.h>
-#include <retry.h>
+#include <sys/param.h>
+#include <sys/stat.h>
+#include <sys/types.h>
+#include <syslog.h>
+#include <unistd.h>
 
-#include "prot.h"
-#include "tls.h"
-#include "util.h"
-#include "global.h"
 #include "libconfig.h"
 #include "xmalloc.h"
 #include "xstrlcat.h"
 #include "xstrlcpy.h"
-#include "sieve_interface.h"
-
-#include "codes.h"
-#include "actions.h"
-#include "scripttest.h"
-
-#include "sync_log.h"
+#include "imap/global.h"
+#include "imap/sync_log.h"
+#include "imap/tls.h"
+#include "imap/version.h"
+#include "timsieved/actions.h"
+#include "timsieved/codes.h"
+#include "timsieved/scripttest.h"
 
 /* after a user has authentication, our current directory is their Sieve 
    directory! */
@@ -160,15 +151,15 @@ int actions_setuser(const char *userid)
  *
  */
 
-static int scriptname_valid(mystring_t *name)
+static int scriptname_valid(const struct buf *name)
 {
-  int lup;
+  unsigned int lup;
   char *ptr;
 
   /* must be at least one character long */
   if (name->len < 1) return TIMSIEVE_FAIL;
 
-  ptr=string_DATAPTR(name);
+  ptr = name->s;
 
   for (lup=0;lup<name->len;lup++)
   {
@@ -186,15 +177,25 @@ int capabilities(struct protstream *conn, sasl_conn_t *saslconn,
     int mechcount;
 
     /* implementation */
-    prot_printf(conn,
-		"\"IMPLEMENTATION\" \"Cyrus timsieved%s %s\"\r\n",
-		config_mupdate_server ? " (Murder)" : "", cyrus_version());
+    if (config_serverinfo == IMAP_ENUM_SERVERINFO_ON) {
+	prot_printf(conn,
+		    "\"IMPLEMENTATION\" \"Cyrus timsieved%s %s\"\r\n",
+		    config_mupdate_server ? " (Murder)" : "", cyrus_version());
+    } else if (config_serverinfo == IMAP_ENUM_SERVERINFO_MIN) {
+	prot_printf(conn,
+		    "\"IMPLEMENTATION\" \"Cyrus timsieved%s\"\r\n",
+		    config_mupdate_server ? " (Murder)" : "");
+    } else {
+    /* IMAP_ENUM_SERVERINFO_OFF */
+	prot_printf(conn,
+		    "\"IMPLEMENTATION\" \"ManageSieve\"\r\n");
+    }
     
     /* SASL */
-    if ((!authenticated || sasl_ssf) &&
-	sasl_listmech(saslconn, NULL, 
-		    "\"SASL\" \"", " ", "\"\r\n",
-		    &sasllist,
+    if (!sieved_tls_required && (!authenticated || sasl_ssf) &&
+	sasl_listmech(saslconn, NULL,
+		      "\"SASL\" \"", " ", "\"\r\n",
+		      &sasllist,
 		      NULL, &mechcount) == SASL_OK/* && mechcount > 0*/)
     {
       prot_printf(conn,"%s",sasllist);
@@ -213,7 +214,7 @@ int capabilities(struct protstream *conn, sasl_conn_t *saslconn,
     return TIMSIEVE_OK;
 }
 
-int getscript(struct protstream *conn, mystring_t *name)
+int getscript(struct protstream *conn, const struct buf *name)
 {
   FILE *stream;
   struct stat filestats;	/* returned by stat */
@@ -231,7 +232,7 @@ int getscript(struct protstream *conn, mystring_t *name)
   }
 
 
-  snprintf(path, 1023, "%s.script", string_DATAPTR(name));
+  snprintf(path, 1023, "%s.script", name->s);
 
   result = stat(path, &filestats);
   if (result != 0) {
@@ -312,13 +313,13 @@ static int countscripts(char *name)
 
 
 /* save name as a sieve script */
-int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
-	      int verify_only)
+int putscript(struct protstream *conn, const struct buf *name,
+	      const struct buf *data, int verify_only)
 {
   FILE *stream;
-  char *dataptr;
+  const char *dataptr;
   char *errstr;
-  int i;
+  unsigned int i;
   int last_was_r = 0;
   int result;
   char path[1024], p2[1024];
@@ -340,7 +341,7 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
       /* see if this would put the user over quota */
       maxscripts = config_getint(IMAPOPT_SIEVE_MAXSCRIPTS);
 
-      if (countscripts(string_DATAPTR(name))+1 > maxscripts)
+      if (countscripts(name->s)+1 > maxscripts)
       {
 	  prot_printf(conn,
 		      "NO (QUOTA/MAXSCRIPTS) \"You are only allowed %d scripts on this server\"\r\n",
@@ -348,7 +349,7 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
 	  return TIMSIEVE_FAIL;
       }
 
-      snprintf(path, 1023, "%s.script.NEW", string_DATAPTR(name));
+      snprintf(path, 1023, "%s.script.NEW", name->s);
 
       stream = fopen(path, "w+");
   }
@@ -360,7 +361,7 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
       return TIMSIEVE_NOEXIST;
   }
 
-  dataptr = string_DATAPTR(data);
+  dataptr = data->s;
 
   /* copy data to file - replacing any lone \r or \n with the
    * \r\n pair so notify messages are SMTP compatible */ 
@@ -386,14 +387,14 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
   result = is_script_parsable(stream, &errstr, &s);
 
   if (result != TIMSIEVE_OK) {
-      if (errstr && *errstr) { 
-	  prot_printf(conn, "NO {" SIZE_T_FMT "}\r\n%s\r\n",
-		      strlen(errstr), errstr);
-	  free(errstr);
+      if (errstr && *errstr) {
+	  prot_printf(conn, "NO ");
+	  prot_printstring(conn, errstr);
+	  prot_printf(conn, "\r\n");
       } else {
-	  if (errstr) free(errstr);
 	  prot_printf(conn, "NO \"parse failed\"\r\n");
       }
+      free(errstr);
       fclose(stream);
       unlink(path);
       return result;
@@ -415,7 +416,7 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
       }
 
       /* Now, open the new file */
-      snprintf(bc_path, 1023, "%s.bc.NEW", string_DATAPTR(name));
+      snprintf(bc_path, 1023, "%s.bc.NEW", name->s);
       fd = open(bc_path, O_CREAT | O_TRUNC | O_WRONLY, 0600);
       if(fd < 0) {
 	  unlink(path);
@@ -442,8 +443,8 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
       close(fd);
 
       /* Now, rename! */
-      snprintf(p2, 1023, "%s.script", string_DATAPTR(name));
-      snprintf(bc_p2, 1023, "%s.bc", string_DATAPTR(name));
+      snprintf(p2, 1023, "%s.script", name->s);
+      snprintf(bc_p2, 1023, "%s.bc", name->s);
       rename(path, p2);
       rename(bc_path, bc_p2);
 
@@ -460,6 +461,12 @@ int putscript(struct protstream *conn, mystring_t *name, mystring_t *data,
 static int deleteactive(struct protstream *conn)
 {
     if (unlink("defaultbc") != 0) {
+	if (errno == ENOENT) {
+	    /* RFC 5804, 2.8 SETACTIVE Command:
+	     * Disabling an active script when there is no script active is
+	     * not an error and MUST result in an OK reply. */
+	    return TIMSIEVE_OK;
+	}
 	prot_printf(conn,"NO \"Unable to unlink active script\"\r\n");
 	return TIMSIEVE_FAIL;
     }
@@ -474,15 +481,20 @@ static int isactive(char *name)
 {
     char filename[1024];
     char activelink[1024];
+    ssize_t link_len;
 
     snprintf(filename, 1023, "%s.bc", name);
     memset(activelink, 0, sizeof(activelink));
-    if ((readlink("defaultbc", activelink, sizeof(activelink)-1) < 0) && 
-	(errno != ENOENT)) 
+
+    link_len = readlink("defaultbc", activelink, sizeof(activelink)-1);
+    if (link_len == -1)
     {
-	syslog(LOG_ERR, "readlink(defaultbc): %m");
+	if (errno != ENOENT)
+	    syslog(LOG_ERR, "readlink(defaultbc): %m");
+
 	return FALSE;
     }
+    activelink[link_len] = '\0';
 
     if (!strcmp(filename, activelink)) {
 	return TRUE;
@@ -492,7 +504,7 @@ static int isactive(char *name)
 }
 
 /* delete a sieve script */
-int deletescript(struct protstream *conn, mystring_t *name)
+int deletescript(struct protstream *conn, const struct buf *name)
 {
   int result;
   char path[1024];
@@ -504,9 +516,9 @@ int deletescript(struct protstream *conn, mystring_t *name)
       return result;
   }
 
-  snprintf(path, 1023, "%s.script", string_DATAPTR(name));
+  snprintf(path, 1023, "%s.script", name->s);
 
-  if (isactive(string_DATAPTR(name))) {
+  if (isactive(name->s)) {
     prot_printf(conn, "NO (ACTIVE) \"Active script cannot be deleted\"\r\n");
     return TIMSIEVE_FAIL;
   }
@@ -515,13 +527,13 @@ int deletescript(struct protstream *conn, mystring_t *name)
 
   if (result != 0) {
       if (result == ENOENT)
-          prot_printf(conn, "NO (NONEXISTENT) \"Script %s does not exist.\"\r\n", string_DATAPTR(name));
+	  prot_printf(conn, "NO (NONEXISTENT) \"Script %s does not exist.\"\r\n", name->s);
       else
           prot_printf(conn,"NO \"Error deleting script\"\r\n");
       return TIMSIEVE_FAIL;
   }
 
-  snprintf(path, 1023, "%s.bc", string_DATAPTR(name));
+  snprintf(path, 1023, "%s.bc", name->s);
 
   result = unlink(path);
 
@@ -558,12 +570,9 @@ int listscripts(struct protstream *conn)
 	{
 	    if (strcmp(dir->d_name + (length - 7), ".script")==0)
 	    {
-		char *namewo=(char *) xmalloc(length-6);
+		char *namewo = xstrndup(dir->d_name, length-7);
 
-		memcpy(namewo, dir->d_name, length-7);
-		namewo[length-7]='\0';
-
-		if (isactive(namewo)==TRUE)
+		if (isactive(namewo))
 		    prot_printf(conn,"\"%s\" ACTIVE\r\n", namewo);
 		else
 		    prot_printf(conn,"\"%s\"\r\n", namewo);
@@ -576,7 +585,7 @@ int listscripts(struct protstream *conn)
     closedir(dp);
 
     prot_printf(conn,"OK\r\n");
-
+  
     return TIMSIEVE_OK;
 }
 
@@ -601,13 +610,13 @@ static int exists(char *str)
 
 /* set the sieve script 'name' to be the active script */
 
-int setactive(struct protstream *conn, mystring_t *name)
+int setactive(struct protstream *conn, const struct buf *name)
 {
     int result;
     char filename[1024];
 
     /* if string name is empty, disable active script */
-    if (!strlen(string_DATAPTR(name))) {
+    if (!name->len) {
 	if (deleteactive(conn) != TIMSIEVE_OK)
 	    return TIMSIEVE_FAIL;
 
@@ -622,20 +631,20 @@ int setactive(struct protstream *conn, mystring_t *name)
 	return result;
     }
 
-    if (exists(string_DATAPTR(name))==FALSE)
+    if (exists(name->s)==FALSE)
     {
 	prot_printf(conn,"NO (NONEXISTENT) \"Script does not exist\"\r\n");
 	return TIMSIEVE_NOEXIST;
     }
 
     /* if script already is the active one just say ok */
-    if (isactive(string_DATAPTR(name))==TRUE) {
+    if (isactive(name->s)==TRUE) {
 	prot_printf(conn,"OK\r\n");
-	return TIMSIEVE_OK;  
+	return TIMSIEVE_OK;
     }
 
     /* get the name of the active sieve script */
-    snprintf(filename, sizeof(filename), "%s.bc", string_DATAPTR(name));
+    snprintf(filename, sizeof(filename), "%s.bc", name->s);
 
     /* ok we want to do this atomically so let's
        - make <activesieve>.NEW as a hard link
@@ -661,7 +670,7 @@ int setactive(struct protstream *conn, mystring_t *name)
     return TIMSIEVE_OK;
 }
 
-int cmd_havespace(struct protstream *conn, mystring_t *sieve_name, unsigned long num)
+int cmd_havespace(struct protstream *conn, const struct buf *sieve_name, unsigned long num)
 {
     int result;
     int maxscripts;
@@ -690,7 +699,7 @@ int cmd_havespace(struct protstream *conn, mystring_t *sieve_name, unsigned long
     /* see if this would put the user over quota */
     maxscripts = config_getint(IMAPOPT_SIEVE_MAXSCRIPTS);
 
-    if (countscripts(string_DATAPTR(sieve_name))+1 > maxscripts)
+    if (countscripts(sieve_name->s)+1 > maxscripts)
     {
 	prot_printf(conn,
 		    "NO (QUOTA/MAXSCRIPTS) \"You are only allowed %d scripts on this server\"\r\n",
